Working Line Follower
More work required, but basic line follower is working

+++ libraries/AP_HAL_SITL/SITL_State.h
@@ -84,7 +84,7 @@ public:
     uint16_t current_pin_value;  // pin 12
     uint16_t voltage2_pin_value;  // pin 15
     uint16_t current2_pin_value;  // pin 14
-    int ir_pin_value;       // pin 16
+    float ir_pin_value;       // pin 16
 
     // paths for UART devices
     const char *_uart_path[7] {
